requirements.txt
# requirements.txt
# CodeVerdict - Where AI Code Stands Trial ⚖️

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ML & Experiment Tracking
mlflow==2.9.2
langchain==0.0.354
langsmith==0.0.66

# LLM Providers
openai==1.3.9
anthropic==0.7.10
transformers==4.35.2
torch==2.1.1

# Evaluation & Analysis
arize-phoenix==3.0.1
evals==0.3.0
rouge-score==0.1.2
evaluate==0.4.1

# Orchestration
prefect==2.14.5
prefect-docker==0.3.4

# Data & Annotation
argilla==1.15.0
pandas==2.1.3
numpy==1.24.3
sqlalchemy==2.0.23

# Code Analysis
pylint==3.0.3
black==23.11.0
flake8==6.1.0
bandit==1.7.5

# Utilities
click==8.1.7
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.2

# Security
cryptography==41.0.7
python-jose==3.3.0
passlib==1.7.4